requirements/requirements-dev.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
loguru

# dev packages
rich 
openai==0.27.2
tiktoken 
wandb 
langchain 
unstructured 
tabulate 
pdf2image 
chromadb==0.4.18
feedparser
feedsearch
newspaper3k
pandas==2.0.3
transformers==4.35.2
torch==2.1.1